split up modis.py
