+++ modape/modis/__init__.py
@@ -1,0 +1,17 @@
+'''
+MODAPE: Modis Data Assimilation and Processing Engine
+MODIS submodule
+
+This submodule contains all functions and classes used in the MODIS
+processing chain.
+
+Author: Valentin Pesendorfer, September 2019
+e-mail: valentin.pesendorfer@wfp.org
+License: MIT
+'''
+from .collect import ModisRawH5
+from .download import ModisQuery
+from .smooth import ModisSmoothH5
+from .window import ModisMosaic, modis_tiles
+
+__all__ = ['ModisRawH5', 'ModisQuery', 'ModisSmoothH5', 'ModisMosaic', 'modis_tiles']
